Moved background statistics recording into separate class
